Electron 5 compatibility
The nodeIntegration flag in BrowserWindow webPreferences had a
default value of true in earlier versions of Electron. The default
value has changed to false in Electron 5. A value of true (i.e.
the presence of the Node environment) is necessary for the test
runner to run as it is currently written.